Remove PostSettingsModal and create seperate modals
- Removes the generic 'PostSettingsModal'
- Refactors the SnippetPreviewModal to the GooglePreviewModal
  -> Make it a functional component
  -> Base it on the WordPress modal
- Create a new SettingsModal
  -> This modal accepts a title prop which is used for the title of the modal and the content of the button in the sidebar
  -> The content of the modal is provided via the children prop
- Create a new TwitterPreviewModal
- Create a new FacebookPreviewModal
- Adapt CSS for the new Modal
- Add the newly created modals to the SidebarFill and remove the PostSettingsModal
